refactor: move component on a separate file
